Backend/requirements.txt
asgiref==3.9.1
black==25.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
Django==5.2.4
django-cors-headers==4.7.0
django-crontab==0.7.1
django-filter==25.1
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.0
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
drf-yasg==1.21.10
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Markdown==3.8.2
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2==2.9.12
psycopg2-binary==2.9.12
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
